+++ src/validate/message.schema.ts
@@ -122,6 +122,32 @@ export const mediaMessageSchema: JSONSchema7 = {
   required: ['number', 'mediatype'],
 };
 
+export const ptvMessageSchema: JSONSchema7 = {
+  $id: v4(),
+  type: 'object',
+  properties: {
+    number: { ...numberDefinition },
+    video: { type: 'string' },
+    delay: {
+      type: 'integer',
+      description: 'Enter a value in milliseconds',
+    },
+    quoted: { ...quotedOptionsSchema },
+    everyOne: { type: 'boolean', enum: [true, false] },
+    mentioned: {
+      type: 'array',
+      minItems: 1,
+      uniqueItems: true,
+      items: {
+        type: 'string',
+        pattern: '^\\d+',
+        description: '"mentioned" must be an array of numeric strings',
+      },
+    },
+  },
+  required: ['number'],
+};
+
 export const audioMessageSchema: JSONSchema7 = {
   $id: v4(),
   type: 'object',