+++ CHANGELOG.md
@@ -20,6 +20,7 @@
 * Add indexes to improve performance in Evolution
 * Add logical or permanent message deletion based on env config
 * Add support for fetching multiple instances by key
+* Update instance.controller.ts to filter by instanceName
 
 # 2.1.2 (2024-10-06 10:09)